code review fixing for PR #17
