+++ apps/web/package.json
@@ -50,6 +50,7 @@
     "simple-icons": "^15.22.0",
     "sonner": "^2.0.7",
     "tailwind-merge": "^3.4.0",
+    "uuid": "^13.0.0",
     "vaul": "^1.1.2",
     "zod": "3.23.8",
     "zustand": "^5.0.9"
