+++ apps/web/src/app/(main)/dashboard/products/_components/product-card-view.tsx
@@ -3,11 +3,9 @@
 import { useState } from "react";
 
 import Image from "next/image";
-import { useRouter } from "next/navigation";
 
 import { formatDistanceToNow } from "date-fns";
 import { MoreVertical, Pencil, RefreshCw, Trash2, TrendingDown, TrendingUp } from "lucide-react";
-import { toast } from "sonner";
 
 import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
@@ -23,45 +21,18 @@ import { DeleteProductDialog } from "./delete-product-dialog";
 import { EditProductDialog } from "./edit-product-dialog";
 import { MiniPriceChart } from "./mini-price-chart";
 import type { ProductWithStats } from "./products-view";
+import { useCheckPrice } from "./use-check-price";
 
 interface ProductCardViewProps {
   products: ProductWithStats[];
 }
 
 export function ProductCardView({ products }: ProductCardViewProps) {
-  const router = useRouter();
   const [editingProduct, setEditingProduct] = useState<ProductWithStats | null>(null);
   const [isEditDialogOpen, setIsEditDialogOpen] = useState(false);
   const [deletingProduct, setDeletingProduct] = useState<ProductWithStats | null>(null);
   const [isDeleteDialogOpen, setIsDeleteDialogOpen] = useState(false);
-  const [checkingPriceId, setCheckingPriceId] = useState<string | null>(null);
-
-  const handleCheckPrice = async (product: ProductWithStats) => {
-    setCheckingPriceId(product.id);
-    try {
-      const response = await fetch(`/api/products/${product.id}/check-price`, {
-        method: "POST",
-      });
-      const data = await response.json();
-
-      if (response.ok && data.success) {
-        toast.success("Price check started", {
-          description: "The price will be updated shortly.",
-        });
-        router.refresh();
-      } else {
-        toast.error("Failed to check price", {
-          description: data.error || "An unexpected error occurred",
-        });
-      }
-    } catch {
-      toast.error("Failed to check price", {
-        description: "Network error. Please try again.",
-      });
-    } finally {
-      setCheckingPriceId(null);
-    }
-  };
+  const { handleCheckPrice, checkingPriceId } = useCheckPrice();
 
   const formatPrice = (cents: number, currency: string) => {
     return new Intl.NumberFormat("en-US", {
@@ -119,7 +90,7 @@ export function ProductCardView({ products }: ProductCardViewProps) {
                       </DropdownMenuTrigger>
                       <DropdownMenuContent align="end">
                         <DropdownMenuItem
-                          onClick={() => handleCheckPrice(product)}
+                          onClick={() => handleCheckPrice(product.id)}
                           disabled={checkingPriceId === product.id}
                         >
                           <RefreshCw className={`mr-2 size-4 ${checkingPriceId === product.id ? "animate-spin" : ""}`} />
